Share Material 3 theme across popup surfaces
