+++ src/hooks/M3Theme.test.js
@@ -1,0 +1,37 @@
+/* eslint-disable testing-library/no-container, testing-library/no-unnecessary-act */
+import { act } from "react";
+import { createRoot } from "react-dom/client";
+import M3Theme from "./M3Theme";
+
+globalThis.IS_REACT_ACT_ENVIRONMENT = true;
+
+jest.mock("./ColorMode", () => ({
+  useDarkMode: () => ({ darkMode: "auto" }),
+}));
+jest.mock("./SystemColorScheme", () => ({
+  useSystemDarkPreference: () => true,
+}));
+
+test("scopes the resolved Material 3 palette to its root", () => {
+  const container = document.createElement("div");
+  document.body.appendChild(container);
+  const root = createRoot(container);
+
+  act(() => {
+    root.render(
+      <M3Theme>
+        <span>content</span>
+      </M3Theme>
+    );
+  });
+
+  const themeRoot = container.querySelector(".kt-m3-root");
+  expect(themeRoot.dataset.theme).toBe("dark");
+  expect(themeRoot.style.getPropertyValue("--kt-pri")).toBe("#A8C7FA");
+  expect(themeRoot.querySelector("style").textContent).toContain(
+    "@keyframes kt-m3-pop"
+  );
+
+  act(() => root.unmount());
+  container.remove();
+});
