+++ src/views/Popup/PopupTheme.js
@@ -1,0 +1,1 @@
+export { default } from "../../hooks/M3Theme";
